Ok registration does indeed add info to the database
Now it just needs to go back to the home page and be like you registered yay! With confetti and fanfare, otherwise this may as well just be a fake website I mean come on! Anyways there is 0 chance I set up JavaScript on the home page that checks whether or not you are logged in (right now) so that can wait till later. I still need to make sure the logging in part works too